Fix tests
Some tests seem to be broken against the current gradle master due
to changes in the field order.

This commit fixes those cases, but maybe we should make the tests
generally more resilient against field order changes.

We keep this around in a branch until we switch to the future
grapql-java v12.0.